+++ Sources/Sentry/include/SentryNSTimerWrapper.h
@@ -8,6 +8,12 @@ NS_ASSUME_NONNULL_BEGIN
                                     repeats:(BOOL)repeats
                                       block:(void (^)(NSTimer *timer))block;
 
+- (NSTimer *)scheduledTimerWithTimeInterval:(NSTimeInterval)ti
+                                     target:(id)aTarget
+                                   selector:(SEL)aSelector
+                                   userInfo:(nullable id)userInfo
+                                    repeats:(BOOL)yesOrNo;
+
 @end
 
 NS_ASSUME_NONNULL_END