+++ Tests/SentryTests/Integrations/Breadcrumbs/SentryAutoBreadcrumbTrackingIntegrationTests.swift
@@ -93,9 +93,9 @@ class SentryAutoBreadcrumbTrackingIntegrationTests: XCTestCase {
 
 private class SentryTestBreadcrumbTracker: SentryBreadcrumbTracker {
     
-    let startInvocations = Invocations<Void>()
-    override func start() {
-        startInvocations.record(Void())
+    let startInvocations = Invocations<SentryBreadcrumbDelegate>()
+    override func start(with delegate: SentryBreadcrumbDelegate) {
+        startInvocations.record(delegate)
     }
     
     let startSwizzleInvocations = Invocations<Void>()
@@ -107,8 +107,8 @@ private class SentryTestBreadcrumbTracker: SentryBreadcrumbTracker {
 
 private class SentryTestSystemEventBreadcrumbs: SentrySystemEventBreadcrumbs {
     
-    let startWithdelegateInvocations = Invocations<SentrySystemEventBreadcrumbsDelegate>()
-    override func start(with delegate: SentrySystemEventBreadcrumbsDelegate) {
+    let startWithdelegateInvocations = Invocations<SentryBreadcrumbDelegate>()
+    override func start(with delegate: SentryBreadcrumbDelegate) {
         startWithdelegateInvocations.record(delegate)
     }
 }