+++ CHANGELOG.md
@@ -2,11 +2,23 @@
 
 ## Unreleased
 
+### Fixes 
+
+- Ensure the current GPU frame rate is always reported for concurrent transaction profiling metrics (#2929)
+
+## 8.5.0
+
+### Features
+
+- feat: Core data operation in the main thread (#2879)
+
 ### Fixes
 
 - Crash when serializing invalid objects (#2858)
 - Convert remaining usages of `sprintf` to `snprintf` (#2866)
 - Don't send screenshots with either width or height of 0 (#2876)
+- Don't send screenshots with either width or height of 0 (#2876)
+- GPU frame alignment with stack traces in profiles (#2856)
 
 ## 8.4.0